Rename Archiver to Archive.
